make keycloak endpoint work
