+++ src/main/java/org/restfull/api/UserResourceProvider.java
@@ -5,13 +5,6 @@ package org.restfull.api;
 import org.keycloak.models.KeycloakSession;
 import org.keycloak.services.resource.RealmResourceProvider;
 
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.Path;
-
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.core.MediaType;
-import jakarta.ws.rs.core.Response;
-
 public class UserResourceProvider implements RealmResourceProvider {
     private KeycloakSession session;
 
@@ -21,25 +14,9 @@ public class UserResourceProvider implements RealmResourceProvider {
 
     @Override
     public Object getResource() {
-        return new UserResource();
+        return new UserResource(session);
     }
 
-    @GET @Path("")
-    @Produces("text/plain; charset=utf-8")
-    public String get() {
-        String name = session.getContext().getRealm().getDisplayName();
-        if (name == null) {
-            name = session.getContext().getRealm().getName();
-        }
-        return "Hello " + name;
-    }
-
-    @GET
-    @Path("/user")
-    @Produces(MediaType.APPLICATION_JSON)
-    public Response getCustomData() {
-        return Response.ok("Custom user data").build();
-    }
 
     @Override
     public void close() {
